+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Fix` *Tunes API* - Tune config now passed to constructor under `config` property
 - `Fix` *Types* - Add common type for internal and external Tools configuration
 - `Fix` — Block's destroy method is called on block deletion
+- `Fix` - Fix jump to the button of editor zone on CBS
 
 ### 2.22.1